#1123 chore(bash): delete duplicated buildBwrapArgs, use canonical sandbox implementation (extracted from #1166)
`packages/boring-bash/src/agent/runtime/buildBwrapArgs.ts` was a 141-line
byte-identical copy of the canonical
`packages/boring-sandbox/src/providers/bwrap/buildBwrapArgs.ts` — a duplicated
security-critical argument builder kept in sync by hand.

Re-verified on current main before deleting:
- `diff` between the two files is empty (byte-identical, 141 lines each).
- The `boring-bash` value exports (`buildBwrapArgs`, `BWRAP_TIMEOUT_SECONDS`,
  `KILL_GRACE_SECONDS`, `RO_BIND_DIRS`, `RO_BIND_TRY_DIRS`) have zero importers
  repo-wide. `packages/agent/host/sandbox.ts` and
  `packages/agent/test-host/sandbox.ts` already import the canonical builder
  from `@hachej/boring-sandbox/providers/bwrap`.

The only in-repo consumer was the type-only
`import type { BwrapArgsOptions }` in `agent/runtime/types.ts`. boring-bash must
not import boring-sandbox (enforced by `check:invariants`), and the builder is
injected through `RuntimeHostOperations.buildBwrapArgs`, so the options
interface moves verbatim into `agent/runtime/types.ts` as the structural
contract for that injection point and is re-exported from the same barrel.

Purely subtractive: no behavior change, no mount-contract code.

Extracted from PR #1166, whose remaining mount-contract substrate is on hold
pending the #1317 reconciliation the owner requested.
