+++ packages/boring-bash/src/agent/runtime/types.ts
@@ -3,8 +3,29 @@ import type {
   Sandbox,
   Workspace,
 } from '@hachej/boring-agent/shared'
-import type { BwrapArgsOptions } from './buildBwrapArgs'
 import type { WorkspacePythonEnvOptions } from './workspacePythonEnv'
+
+/**
+ * Options contract for the injected bwrap args builder. The canonical
+ * implementation (and its validation rules) is
+ * `@hachej/boring-sandbox/providers/bwrap#buildBwrapArgs`; boring-bash never
+ * imports the provider package, so hosts inject the builder through
+ * {@link RuntimeHostOperations}. This interface is the structural mirror of
+ * that builder's options and must stay in sync with it.
+ */
+export interface BwrapArgsOptions {
+  extraArgs?: string[]
+  postWorkspaceArgs?: string[]
+  network?: 'shared' | 'isolated'
+  newSession?: boolean
+  dropAllCapabilities?: boolean
+  /**
+   * Workspace-relative prefixes re-bound readonly on top of the writable
+   * workspace mount, so spawned shells cannot mutate protected paths that the
+   * Operations layer already refuses to mutate.
+   */
+  readonlyPaths?: readonly string[]
+}
 
 export type RuntimeBashStrategy =
   | { kind: 'host'; preserveHostHome?: boolean }
